+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 ### Changed
 - `Pck::configure` now takes a `u16` prescaler instead of a `u8`, and returns `Err(PckError)` if the prescaler cannot be applied.
+- Renamed `serial::ExtU32` to `serial::ExtBpsU32`, in order to not clash with `fugit::ExtU32`.
 
 ## [v0.2.1] 2022-08-08